Rename wemo coordinator module
