+++ homeassistant/components/wemo/sensor.py
@@ -19,8 +19,8 @@ from homeassistant.helpers.entity_platform import AddEntitiesCallback
 from homeassistant.helpers.typing import StateType
 
 from . import async_wemo_dispatcher_connect
+from .coordinator import DeviceCoordinator
 from .entity import WemoEntity
-from .wemo_device import DeviceCoordinator
 
 
 @dataclass(frozen=True)
